+++ default.nix
@@ -11,12 +11,8 @@ in
 with super.lib;
 
 (foldl' (flip extends) (_: super) [
-
-  (import localLib.fetchNixPkgsStackage)
   (import localLib.fetchNixPkgsLibQuixoftic)
-
   (import ./nix/overlays/lib.nix)
   (import ./nix/overlays/haskell-lib.nix)
   (import ./nix/overlays/haskell-overrides.nix)
-
 ]) self