+++ nix/jobsets/default.nix
@@ -15,14 +15,6 @@ let
     emailresponsible = false;
   };
 
-  ## nixpkgs-stackage wants a <nixpkgs> path so that it can import
-  ## Haskell functions that nixpkgs doesn't explicitly export.
-  nixpkgs-src = builtins.fromJSON (builtins.readFile ../nixpkgs-src.json);
-  nixpkgs-spec = {
-    url = "https://github.com/${nixpkgs-src.owner}/${nixpkgs-src.repo}.git";
-    rev = "${nixpkgs-src.rev}";
-  };
-
   pkgs = import nixpkgs {};
 
   defaultSettings = {
@@ -37,25 +29,13 @@ let
     nixexprinput = "hpio";
     description = "hpio";
     inputs = {
-
-      ## Note: the nixpkgs input here is for nixpkgs-stackage's
-      ## <nixpkgs>. It is not used by hpio.
-      nixpkgs = mkFetchGithub "${nixpkgs-spec.url} ${nixpkgs-spec.rev}";
-
       hpio = mkFetchGithub "${hpioUri} master";
-
     };
   };
 
-  mkChannelAlt = hpioBranch: nixpkgsRev: nixpkgsStackageRev: {
+  mkChannelAlt = hpioBranch: nixpkgsRev: {
     inputs = {
-
-      ## Note: the nixpkgs input here is for nixpkgs-stackage's
-      ## <nixpkgs>. It is not used by hpio.
-      nixpkgs = mkFetchGithub "${nixpkgs-spec.url} ${nixpkgs-spec.rev}";
-
       nixpkgs_override = mkFetchGithub "https://github.com/NixOS/nixpkgs-channels.git ${nixpkgsRev}";
-      nixpkgs_stackage_override = mkFetchGithub "https://github.com/typeable/nixpkgs-stackage.git ${nixpkgsStackageRev}";
       hpio = mkFetchGithub "${hpioUri} ${hpioBranch}";
 
     };
@@ -65,17 +45,12 @@ let
   ## "next" builds; these are expected to fail from time to time and
   ## don't run as often. They also build from nixpkgs and not
   ## nixpkgs-channels as we always want to be testing the latest,
-  ## greatest GHC pre-releases. (We don't bother overriding
-  ## nixpkgs-stackage because it's not used in these evaluations.)
+  ## greatest GHC pre-releases.
 
   mkNext = hpioBranch: nixpkgsRev: {
     nixexprpath = "nix/jobsets/next.nix";
     checkinterval = 60 * 60 * 24;
     inputs = {
-
-      ## Note: does not depend on nixpkgs because we don't use
-      ## nixpkgs-stackage for these builds.
-
       nixpkgs_override = mkFetchGithub "https://github.com/NixOS/nixpkgs.git ${nixpkgsRev}";
       hpio = mkFetchGithub "${hpioUri} ${hpioBranch}";
 
@@ -84,7 +59,7 @@ let
 
   mainJobsets = with pkgs.lib; mapAttrs (name: settings: defaultSettings // settings) (rec {
     master = {};
-    nixpkgs-unstable = mkChannelAlt "master" "nixpkgs-unstable" "master";
+    nixpkgs-unstable = mkChannelAlt "master" "nixpkgs-unstable";
     next-ghc = mkNext "master" "master";
   });