+++ nix/jobsets/release.nix
@@ -1,9 +1,7 @@
 ## The full set of packages we build/test, both on Hydra and for more
 ## extensive interactive development and testing. This file will
 ## create Hydra-style jobs for hpio built against a fixed Nixpkgs
-## Haskell package set; plus, via nixpkgs-stackage, all of the
-## Stackage LTS package sets we support (for which Nixpkgs has a
-## compiler, anyway).
+## Haskell package set.
 
 let
 
@@ -26,7 +24,6 @@ with import (fixedNixPkgs + "/pkgs/top-level/release-lib.nix") {
 let
 
   jobs = {
-
     nixpkgs = pkgs.releaseTools.aggregate {
       name = "nixpkgs";
       meta.description = "hpio built against nixpkgs haskellPackages";
@@ -37,43 +34,11 @@ let
         haskellPackages.hpio.aarch64-linux
       ];
     };
-
-    lts-9 = pkgs.releaseTools.aggregate {
-      name = "lts-9";
-      meta.description = "hpio built against Stackage LTS 9 package set";
-      meta.maintainer = lib.maintainers.dhess;
-      constituents = with jobs; [
-        lts9Packages.hpio.x86_64-linux
-      ];
-    };
-
-    lts-11 = pkgs.releaseTools.aggregate {
-      name = "lts-11";
-      meta.description = "hpio built against Stackage LTS 11 package set";
-      meta.maintainer = lib.maintainers.dhess;
-      constituents = with jobs; [
-        lts11Packages.hpio.x86_64-linux
-      ];
-    };
-
-    lts-12 = pkgs.releaseTools.aggregate {
-      name = "lts-12";
-      meta.description = "hpio built against Stackage LTS 12 package set";
-      meta.maintainer = lib.maintainers.dhess;
-      constituents = with jobs; [
-        lts12Packages.hpio.x86_64-linux
-      ];
-    };
-
   } // (mapTestOn ({
-
     haskellPackages = packagePlatforms pkgs.haskellPackages;
-    lts9Packages = packagePlatforms pkgs.lts9Packages;
-    lts11Packages = packagePlatforms pkgs.lts11Packages;
-    lts12Packages = packagePlatforms pkgs.lts12Packages;
   }));
 
 in
 {
-  inherit (jobs) nixpkgs lts-9 lts-11 lts-12;
+  inherit (jobs) nixpkgs;
 }