+++ nix/lib.nix
@@ -10,14 +10,6 @@ let
       then builtins.trace "Using <nixpkgs_override>" try.value
       else (import ./fetch-github.nix) { jsonSpec = builtins.readFile ./nixpkgs-src.json; };
 
-  fetchNixPkgsStackage =
-  let
-    try = builtins.tryEval <nixpkgs_stackage_override>;
-  in
-    if try.success
-      then builtins.trace "Using <nixpkgs_stackage_override>" try.value
-      else (import ./fetch-github.nix) { jsonSpec = builtins.readFile ./nixpkgs-stackage-src.json; };
-
   fetchNixPkgsLibQuixoftic =
   let
     try = builtins.tryEval <nixpkgs_lib_quixoftic_override>;
@@ -34,7 +26,6 @@ let
 in lib // (rec {
 
   inherit fetchNixPkgs;
-  inherit fetchNixPkgsStackage;
   inherit fetchNixPkgsLibQuixoftic;
 
 })